requirements.txt
selenium>=4.10.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
pypdf>=4.0.0
WeasyPrint>=59.0 